pyproject.toml
[project]
name = "pdf-skimmer-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.13.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.33",
    "langchain-deepseek>=0.0.1",
    "langchain-google-genai>=2.0.9",
    "langchain-openai>=0.3.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "tomli>=2.2.1",
]
